I´ve refactored integration test and moved other files to commons module
