+++ SNK/SNK/Commons/CoreData/CoreDataModels/Information+CoreDataClass.swift
