+++ SNK/SNK/Commons/CoreData/CoreDataModels/Information+CoreDataProperties.swift
