+++ SNK/SNK/Commons/CoreData/Providers/CoreDataProvider.swift
