+++ SNK/SNK/Commons/CoreData/SnkDM.xcdatamodeld/SnkDM.xcdatamodel/Contents
