+++ SNK/SNK/Commons/CoreData/Transformers/Information+Transformer.swift
