+++ SNK/SNK/Commons/CoreData/Transformers/String+Transformer.swift
