+++ SNK/SNK/Commons/Resources/Localizable.xcstrings
