+++ SNK/SNK/Commons/Utils/Errors.swift
